requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.2.0
autobahn==22.12.1
Automat==22.10.0
beautifulsoup4==4.11.1
blog==1.0.1
certifi==2022.9.24
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==2.1.1
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
Django==4.2
django-filter==22.1
django-jalali-date==1.0.1
django-qr-code==3.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
form==0.0.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ippanel==2.0.7
jdatetime==4.1.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
nano==0.10.0
neshan==1.1.1
num2words==0.5.12
Pillow==9.3.0
psycopg2==2.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.4
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.13.0
PyJWT==2.6.0
pymdown-extensions==9.9
pyOpenSSL==23.3.0
pytz==2022.6
PyYAML==6.0
redis==5.0.1
regex==2022.10.31
requests==2.28.1
segno==1.5.2
service-identity==23.1.0
simplejson==3.18.1
six==1.16.0
soupsieve==2.3.2.post1
spinners==0.0.24
sqlparse==0.4.3
suds==1.1.2
termcolor==1.1.0
transaction==3.0.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2023.4
Unidecode==1.3.6
urllib3==1.26.13
wcwidth==0.2.5
zope.interface==5.5.2
